+++ docs/DECISIONS.md
@@ -193,6 +193,7 @@
 - Static HTML top-bar rendering lives in `WorkspaceHTMLTopBarRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates title identity, status, runtime issue pill, Computer Use status, and overflow command markup while staying responsible for whole-workspace shell composition.
 - Static HTML transcript rendering lives in `WorkspaceHTMLTranscriptRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates transcript empty state, context banner, runtime issue panel, review placement, message timeline rows, tool-card row handoff, and composer/model/mode markup while staying responsible only for composing the whole static harness document.
 - Browser state and presentation contracts live in `QuillCodeBrowserSurface.swift`. `BrowserState`, `BrowserSnapshotState`, `BrowserCommentState`, `BrowserSurface`, `BrowserSnapshotSurface`, and `BrowserCommentSurface` evolve together; browser address resolution lives in `WorkspaceBrowserLocationResolver.swift`; and browser history/page/comment state transitions live in `WorkspaceBrowserEngine`. The workspace model owns async fetch orchestration, `lastError`, and top-bar refreshes, while reusable browser state contracts, URL-resolution rules, and pure state transitions stay focused and guarded by parity tests.
+- Workspace browser integration coverage lives in `WorkspaceBrowserIntegrationTests`, not `WorkspaceModelTests`. Browser URL normalization, local HTML snapshots, reachable-page fetches, fetch-failure fallback behavior, history navigation, page comments, and composer-driven browser inspection cross the workspace model, browser state, HTML snapshotting, tool cards, and transcripts; keep those flows together outside the model monolith while pure browser state and rendering coverage stays in `WorkspaceBrowserEngineTests` and `QuillCodeBrowserSurfaceTests`.
 - Review comment payload state and event planning live in `WorkspaceReviewCommentPlanner.swift`, not in `WorkspaceModel.swift`. The planner owns path/text trimming, visible review-file validation, line-range normalization, line-kind checks, summary copy, and payload encoding. The workspace model owns only selected-thread validation, event append, persistence, and top-bar refresh.
 - Review action tool calls live in `WorkspaceReviewActionToolCallPlanner.swift`, not as extensions on review surface values or inline in `WorkspaceModel.swift`. The planner owns the canonical file/hunk stage and restore `host.git.*` argument shape, including hunk patch fallback behavior. The workspace model owns only review action execution, follow-up diff refresh, transcript persistence, and top-bar state.
 - Production Swift sources should avoid `try!`, `as!`, and force unwraps. Startup, OAuth sign-in, local project scanning, and extension discovery must fail through typed errors or ignored unsafe records rather than crashing the process. `ProjectExtensionManifestLoader` normalizes extension directories by safe path components and skips unsafe custom directories without blocking later valid directories.
